Move immutable type definitions to their own file
